pyproject.toml
[project]
name = "roulette-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.3",
    "python-socketio[client]>=5.16.0",
    "rich>=14.2.0",
]
